+++ src/main/java/com/shootemup/g53/model/element/Spaceship.java
@@ -1,15 +1,12 @@
 package com.shootemup.g53.model.element;
 
-import com.shootemup.g53.controller.firing.FiringStrategy;
-import com.shootemup.g53.controller.game.BulletPoolController;
-import com.shootemup.g53.controller.movement.MovementStrategy;
 import com.shootemup.g53.model.util.Position;
 import com.shootemup.g53.model.util.objectpool.PoolableObject;
 
 public class Spaceship extends MovableElement {
     private int height;
     private int health;
-    public Spaceship(Position position, int height,int health, String color, int speed) {
+    public Spaceship(Position position, int height, int health, String color, double speed) {
         super(position, color, speed);
         this.height = height;
         this.health = health;
